refactor(plugins): move PluginMap and friends to a different file
